requirements.txt
streamlit>=1.24.0
sqlalchemy>=2.0
pandas>=1.5
scikit-learn>=1.2
matplotlib>=3.6
seaborn>=0.12
requests>=2.28
sentence-transformers>=2.2
torch>=2.0
scipy>=1.10
pyyaml>=6.0
shap>=0.42
statsmodels>=0.14
tqdm>=4.64
numpy>=1.23
